% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Messages.R
\name{CollapseTexts}
\alias{CollapseTexts}
\title{CollapseTexts}
\usage{
CollapseTexts(..., collapse = " ")
}
\arguments{
\item{...}{Text to be collapsed}

\item{collapse}{String to be used for concatenating texts}
}
\description{
CollapseTexts
}
\examples{
CollapseTexts('Item 1', 'Item 2')

}
