% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Messages.R
\name{EndProcess}
\alias{EndProcess}
\title{EndProcess}
\usage{
EndProcess(
  processId = NULL,
  ...,
  collapse = " ",
  .envir = parent.frame(),
  verbose = TRUE
)
}
\arguments{
\item{processId}{Process Id}

\item{...}{Text to be printed}

\item{collapse}{String to be used for concatenating texts}

\item{.envir}{Environment for lookup of variables referenced in the text}

\item{verbose}{Logical indicating to print out messages (TRUE) or not (FALSE)}
}
\description{
EndProcess
}
\examples{
\dontrun{
EndProcess(processId, 'Test')
}
}
