% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAllowedYearRanges.R
\name{GetAllowedYearRanges}
\alias{GetAllowedYearRanges}
\title{GetAllowedYearRanges}
\usage{
GetAllowedYearRanges(data)
}
\arguments{
\item{data}{List of incidence data items}
}
\value{
Named list of vectors of year ranges per data item
}
\description{
Get allowed year ranges for incidence data
}
\examples{
\dontrun{
GetAllowedYearRanges(data)
}

}
