% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDefaultRunContex.R
\name{GetDefaultRunContex}
\alias{GetDefaultRunContex}
\title{GetDefaultRunContex}
\usage{
GetDefaultRunContex()
}
\value{
context list object
}
\description{
Description
}
\examples{
GetDefaultRunContex()

}
