% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetInfoList.R
\name{GetInfoList}
\alias{GetInfoList}
\title{GetInfoList}
\usage{
GetInfoList(context)
}
\arguments{
\item{context}{List of parameters. Required.}
}
\value{
list
}
\description{
Get \code{info} list object.
}
\examples{
\dontrun{
GetInfoList(context)
}

}
