% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetIntervalsFromData.R
\name{GetIntervalsFromData}
\alias{GetIntervalsFromData}
\title{GetIntervalsFromData}
\usage{
GetIntervalsFromData(
  minYear,
  maxYear,
  numIntervals,
  firstIntervalEndYear = NULL
)
}
\arguments{
\item{minYear}{Start year for the diagnosis rate}

\item{maxYear}{End year for the diagnosis rate}

\item{numIntervals}{Number of intervals}

\item{firstIntervalEndYear}{End year of the first interval. Optional. Default = NULL}
}
\value{
Intervals specification as data.table
}
\description{
Get diagnosis rates intervals initialized from allowed year ranges
}
\examples{
\dontrun{
GetIntervalsFromData(minYear, maxYear, numIntervals, firstIntervalEndYear)
}

}
