% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetModelOutputs.R
\name{GetModelOutputs}
\alias{GetModelOutputs}
\title{GetModelOutputs}
\usage{
GetModelOutputs(model, data)
}
\arguments{
\item{model}{model}

\item{data}{data}
}
\value{
list of outputs
}
\description{
Get model outputs
}
\examples{
\dontrun{
GetModelOutputs(model, info, param, data)
}

}
