% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetModelPop.R
\name{GetModelPop}
\alias{GetModelPop}
\title{GetModelPop}
\usage{
GetModelPop(country = "")
}
\arguments{
\item{country}{Country code}
}
\value{
Input data as data.table
}
\description{
Description
}
\examples{
GetModelPop(country = 'NL')
GetModelPop(country = 'PL')
GetModelPop()

}
