% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetObjectDefinition.R
\name{GetObjectDefinition}
\alias{GetObjectDefinition}
\title{GetObjectDefinition}
\usage{
GetObjectDefinition(
  objectName,
  path = system.file("ObjectDefinitions", package = "hivModelling"),
  section = NULL,
  includeFileName = FALSE
)
}
\arguments{
\item{objectName}{Name of the object. Required.}

\item{path}{Path to the object definition file. Optional.
Default = \code{\link{system.file}('ObjectDefinitions',
package = 'hivModelling')}.}

\item{section}{Name of section of the list to return. If not specified, then
the whole specification is returned, otherwise only the specified section.
Optional. Default = \code{NULL}.}

\item{includeFileName}{Logical indicating to include file name in the
returned list as field 'FileName'. Optional. Default = \code{FALSE}.}
}
\value{
List object.
}
\description{
Returns object definition stored as text file.
}
\examples{
\dontrun{
GetObjectDefinition(objectName = 'Model')
GetObjectDefinition(objectName = 'Model', section = 'Name')
GetObjectDefinition(objectName = 'Settings', section = 'Columns')
}

}
