% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetOutputArtifacts.R
\name{GetOutputArtifacts}
\alias{GetOutputArtifacts}
\title{GetOutputArtifacts}
\usage{
GetOutputArtifacts(results)
}
\arguments{
\item{results}{List of results. Required.}
}
\value{
artifacts list object
}
\description{
Creates a set of output artifacts given model output results.
}
\examples{
\dontrun{
GetOutputArtifacts(results)
}

}
