% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetParameterVector.R
\name{GetParameterVector}
\alias{GetParameterVector}
\title{GetParameterVector}
\usage{
GetParameterVector(beta, thetaF)
}
\arguments{
\item{beta}{Numeric vector}

\item{thetaF}{Numeric vector}
}
\value{
numeric vector
}
\description{
Get \code{param} vector initialized from inputs
}
\examples{
\dontrun{
GetParameterVector(beta, thetaF)
}

}
