% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPopulationData.R
\name{GetPopulationData}
\alias{GetPopulationData}
\title{GetPopulationData}
\usage{
GetPopulationData(context, populationSet = NULL)
}
\arguments{
\item{context}{List of parameters. Required.}

\item{populationSet}{Character vector of names of populations to aggregate in the ouptu data set.
Default = \code{NULL}.}
}
\value{
Data set as data.table object
}
\description{
Get data for combinations of populations
}
\examples{
\dontrun{
GetPopulationData(inputData, populationSet = c('pop_0', 'pop_1'))
}

}
