% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetProvSurv96.R
\name{GetProvSurv96}
\alias{GetProvSurv96}
\title{GetProvSurv96}
\usage{
GetProvSurv96(country, noStage, qoppa, modelMinYear, modelNoYears)
}
\arguments{
\item{country}{country}

\item{noStage}{noStage}

\item{qoppa}{qoppa}

\item{modelMinYear}{modelMinYear}

\item{modelNoYears}{modelNoYears}
}
\value{
matrix
}
\description{
Description
}
\examples{
\dontrun{
GetProvSurv96(, noStage, qoppa, modelMinYear, modelNoYears)
}

}
