% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetRunContext.R
\name{GetRunContext}
\alias{GetRunContext}
\title{GetRunContext}
\usage{
GetRunContext(...)
}
\arguments{
\item{...}{Parameters to be merged with default parameters and settings.
Optional.}
}
\value{
context list object
}
\description{
Description
}
\examples{
GetRunContext()

}
