% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonitorFutureJobs.R
\name{MonitorFutureJobs}
\alias{MonitorFutureJobs}
\title{MonitorFutureJobs}
\usage{
MonitorFutureJobs(jobs, statusRefreshRate = 2)
}
\arguments{
\item{jobs}{List of 'future' jobs.}

\item{statusRefreshRate}{Number of seconds to next refresh of job status. Default: 2}
}
\value{
The results of running the futures. This will evaluate any results so if the result is an error
(for example triggered by stop) the error will be triggered and execution will stop.
}
\description{
Monitors the progress of jobs defined with package 'future'.
}
\examples{
\dontrun{
MonitorFutureJobs(jobs)
}

}
