% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerformBootstrapFit.R
\name{PerformBootstrapFit}
\alias{PerformBootstrapFit}
\title{PerformBootstrapFit}
\usage{
PerformBootstrapFit(
  runId,
  context,
  data,
  mainResults,
  maxNoFit = 30,
  ctol = 1e-06,
  ftol = 1e-05,
  verbose = FALSE
)
}
\arguments{
\item{runId}{Bootstrap run index. Must be greater than 0, which is reserved for the main fit.}

\item{context}{List of parameters. Required.}

\item{data}{Input data as data.table. Required.}

\item{mainResults}{Main results}

\item{maxNoFit}{Maximum number of amoeba iterations. Optional. Default = 30.}

\item{ctol}{Minium required deviance in consecutive lambda estimations.
Optional. Default = 1e-6.}

\item{ftol}{Minium required deviance in amoeba calculations. Optional.
Default = 1e-5.}

\item{verbose}{Logical indicating to print detailed info during fitting. Optional.
Default = \code{FALSE}}
}
\value{
Results list object
}
\description{
Perform bootstrap fit
}
\examples{
\dontrun{
PerformBootstrapFit(context, data, maxNoFit = 2, ctol = 1e-6, ftol = 1e-5, verbose = FALSE)
}

}
