% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerformBootstrapFits.R
\name{PerformBootstrapFits}
\alias{PerformBootstrapFits}
\title{PerformBootstrapFits}
\usage{
PerformBootstrapFits(context)
}
\arguments{
\item{context}{List of parameters. Required.}
}
\value{
context list object
}
\description{
Description
}
\examples{
\dontrun{
PerformBootstrapFits(context)
}

}
