% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreprocessInputData.R
\name{PreprocessInputData}
\alias{PreprocessInputData}
\title{PreprocessInputData}
\usage{
PreprocessInputData(inputData, minYear = 1980L, maxYear = year(Sys.time()))
}
\arguments{
\item{inputData}{List of input data. Required.}

\item{minYear}{Model start year.}

\item{maxYear}{Model end year.}
}
\value{
Input data as data.table
}
\description{
Description
}
\examples{
\dontrun{
PreprocessInputData(inputData)
}

}
