% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Messages.R
\name{PrintAlert}
\alias{PrintAlert}
\title{PrintAlert}
\usage{
PrintAlert(
  ...,
  collapse = " ",
  type = "info",
  .envir = parent.frame(),
  verbose = TRUE
)
}
\arguments{
\item{...}{Text to be printed}

\item{collapse}{String to be used for concatenating texts}

\item{type}{Type of alert}

\item{.envir}{Environment for lookup of variables referenced in the text}

\item{verbose}{Logical indicating to print out messages (TRUE) or not (FALSE)}
}
\description{
PrintAlert
}
\examples{
\dontrun{
PrintAlert('Test')
PrintAlert('Test', type = 'success')
PrintAlert('Test', type = 'danger')
PrintAlert('Test', type = 'warning')
PrintAlert('Test', type = 'info')
}
}
