% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunIncidenceModel.R
\name{RunIncidenceModel}
\alias{RunIncidenceModel}
\title{Run incidence model}
\usage{
RunIncidenceModel()
}
\value{
NULL (invisibly)
}
\description{
Executes the incidence model
}
\examples{
RunIncidenceModel()

}
