% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunModels.R
\name{RunModels}
\alias{RunModels}
\title{Run all models}
\usage{
RunModels(argsList = NULL, executionPlan = future::sequential)
}
\arguments{
\item{argsList}{List of arguments to be passed to context (one per model
run). Default = \code{NULL}}

\item{executionPlan}{Execution plan for the job planner (see package
\pkg{future}). Default = \code{\link[future]{sequential}}}
}
\value{
list object containing context, data and model objects
}
\description{
Executes all the specified models
}
\examples{
RunModels()

}
