% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetParameters.R
\name{SetParameters}
\alias{SetParameters}
\title{SetParameters}
\usage{
SetParameters(context)
}
\arguments{
\item{context}{List of parameters. Required.}
}
\value{
context list object
}
\description{
Description
}
\examples{
\dontrun{
SetParameters(context)
}

}
