% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Messages.R
\name{StartProcess}
\alias{StartProcess}
\title{StartProcess}
\usage{
StartProcess(..., collapse = " ", .envir = parent.frame(), verbose = TRUE)
}
\arguments{
\item{...}{Text to be printed}

\item{collapse}{String to be used for concatenating texts}

\item{.envir}{Environment for lookup of variables referenced in the text}

\item{verbose}{Logical indicating to print out messages (TRUE) or not (FALSE)}
}
\description{
StartProcess
}
\examples{
\dontrun{
processId <- StartProcess('Test')
}
}
