% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateInputData.R
\name{ValidateInputData}
\alias{ValidateInputData}
\title{ValidateInputData}
\usage{
ValidateInputData(dt)
}
\arguments{
\item{dt}{Named list of input data sets}
}
\value{
NULL (invisibly)
}
\description{
Performs checks on input data
}
\examples{
\dontrun{
  ValidateInputData(dt)
}

}
