% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateOutputPlots.R
\name{CreateOutputPlots}
\alias{CreateOutputPlots}
\title{CreateOutputPlots}
\usage{
CreateOutputPlots(mainResults, bsResultsList)
}
\arguments{
\item{mainResults}{Output of function \code{\link{PerformMainFit}}. Required.}

\item{bsResultsList}{Output of function \code{\link{PerformBootstrapFits}}. Required.}
}
\value{
list of ggplot2 objects
}
\description{
Creates output plots
}
\examples{
\dontrun{
CreateOutputPlots(mainResults, bsResultsList)
}

}
