% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateParameters.R
\name{EstimateParameters}
\alias{EstimateParameters}
\title{EstimateParameters}
\usage{
EstimateParameters(runType, mainResults = NULL, probSurv1996, param,
  info, data, maxNoFit = 30, ctol = 1e-06, ftol = 1e-05, ...)
}
\arguments{
\item{runType}{String indicating type of run, either 'MAIN' or 'BOOSTRAP'}

\item{mainResults}{Main results}

\item{probSurv1996}{Test}

\item{param}{List of parameters. Required.}

\item{info}{List of parameters. Required.#'}

\item{data}{Input data as data.table. Required.}

\item{maxNoFit}{Maximum number of amoeba iterations. Optional. Default = 30.}

\item{ctol}{Minium required deviance in consecutive lambda estimations.
Optional. Default = 1e-6.}

\item{ftol}{Minium required deviance in amoeba calculations. Optional.
Default = 1e-5.}

\item{...}{Additional arguments passed to amoeba function. Optional.}
}
\value{
NULL (invisibly)
}
\description{
Estimate parameters
}
\examples{
\dontrun{
EstimateParameters(runType, mainResults, probSurv1996, param, info, data, maxNoFit, ctol, ftol)
}

}
