% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitModel.R
\name{FitModel}
\alias{FitModel}
\title{FitModel}
\usage{
FitModel(beta, theta, context, data)
}
\arguments{
\item{beta}{beta}

\item{theta}{theta}

\item{context}{context}

\item{data}{data}
}
\value{
list of outputs
}
\description{
Fit model
}
\examples{
\dontrun{
FitModel(beta, theta, context, data)
}

}
