% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetParamList.R
\name{GetParamList}
\alias{GetParamList}
\title{GetParamList}
\usage{
GetParamList(context, info)
}
\arguments{
\item{context}{List of parameters. Required.}

\item{info}{List of info parameters. Required.}
}
\value{
list
}
\description{
Get \code{param} list object.
}
\examples{
\dontrun{
GetParamList(context, info)
}

}
