% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPopulationData.R
\name{GetPopulationData}
\alias{GetPopulationData}
\title{GetPopulationData}
\usage{
GetPopulationData(context, data)
}
\arguments{
\item{context}{List of parameters. Required.}

\item{data}{Input data as data.table. Required.}
}
\value{
Data set as data.table object
}
\description{
Get data for combinations of populations
}
\examples{
\dontrun{
GetPopulationData(inputData, populationSet = c('pop_0', 'pop_1'))
}

}
