% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetProvSurv96.R
\name{GetProvSurv96}
\alias{GetProvSurv96}
\title{GetProvSurv96}
\usage{
GetProvSurv96(param, info)
}
\arguments{
\item{param}{Parameters list}

\item{info}{Info list}
}
\value{
matrix
}
\description{
Description
}
\examples{
\dontrun{
GetProvSurv96(param, info)
}

}
