% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerformBootstrapFits.R
\name{PerformBootstrapFits}
\alias{PerformBootstrapFits}
\title{PerformBootstrapFits}
\usage{
PerformBootstrapFits(context, data, mainResults)
}
\arguments{
\item{context}{List of parameters. Required.}

\item{data}{Input data as data.table. Required.}

\item{mainResults}{List of nain results. Required.}
}
\value{
Results list object
}
\description{
Description
}
\examples{
\dontrun{
PerformBootstrapFits(context, data, mainResults)
}

}
