% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerformBootstrapFits.R
\name{PerformBootstrapFits}
\alias{PerformBootstrapFits}
\title{PerformBootstrapFits}
\usage{
PerformBootstrapFits(context, data, model)
}
\arguments{
\item{context}{List of parameters. Required.}

\item{data}{Input data as data.table. Required.}

\item{model}{Model data.table object. Required.}
}
\value{
model data.table object
}
\description{
Description
}
\examples{
\dontrun{
PerformBootstrapFits(context, data, model)
}

}
