% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerformBootstrapFits.R
\name{PerformBootstrapFits}
\alias{PerformBootstrapFits}
\title{PerformBootstrapFits}
\usage{
PerformBootstrapFits(
  context,
  data,
  mainResults,
  bsCount = 20,
  maxNoFit = 30,
  ctol = 1e-06,
  ftol = 1e-05,
  executionPlan = future::sequential,
  statusRefreshRate = 2,
  ...
)
}
\arguments{
\item{context}{List of parameters. Required.}

\item{data}{Input data as data.table. Required.}

\item{mainResults}{Main results. Required.}

\item{bsCount}{Count of boostrap iterations. Default = \code{20}}

\item{maxNoFit}{Maximum number of optimization iterations. Optional. Default = 30.}

\item{ctol}{Minium required deviance in consecutive lambda estimations.
Optional. Default = 1e-6.}

\item{ftol}{Minium required deviance in optimization calculations. Optional.
Default = 1e-5.}

\item{executionPlan}{Execution plan for the job planner (see package 'future'). Optional.
Default = future::sequential.}

\item{statusRefreshRate}{Number of seconds to next refresh of job status. Optional. Default = 2.}

\item{...}{Additional arguments passed to \code{\link{PerformBootstrapFit}} function. Optional.}
}
\value{
list object containing context, data and model objects
}
\description{
Performs multiple bootstrap fits
}
\examples{
\dontrun{
PerformBootstrapFits(bsCount, context, data, maxNoFit = 2, ctol = 1e-6, ftol = 1e-5, verbose)
}

}
