% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerformMainFit.R
\name{PerformMainFit}
\alias{PerformMainFit}
\title{PerformMainFit}
\usage{
PerformMainFit(context, data)
}
\arguments{
\item{context}{List of parameters. Required.}

\item{data}{Input data as data.table. Required.}
}
\value{
model data.table object
}
\description{
Description
}
\examples{
\dontrun{
PerformMainFit(context)
}

}
