% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerformMainFit.R
\name{PerformMainFit}
\alias{PerformMainFit}
\title{PerformMainFit}
\usage{
PerformMainFit(
  context,
  data,
  param = NULL,
  info = NULL,
  maxNoFit = 30L,
  ctol = 1e-06,
  ftol = 1e-05,
  algorithm = "NLOPT_LN_BOBYQA",
  verbose = FALSE
)
}
\arguments{
\item{context}{List of parameters. Required.}

\item{data}{Input data as data.table. Required.}

\item{param}{List of parameters. Optional. Default = NULL.}

\item{info}{List of parameters. Optional. Default = NULL.}

\item{maxNoFit}{Maximum number of amoeba iterations. Optional. Default = 30.}

\item{ctol}{Minium required deviance in consecutive lambda estimations.
Optional. Default = 1e-6.}

\item{ftol}{Minium required deviance in amoeba calculations. Optional.
Default = 1e-5.}

\item{algorithm}{Name of optimization algorithm from package \code{nloptr} to use for bootstrap
iterations. Default = 'NLOPT_LN_BOBYQA'}

\item{verbose}{Logical indicating to print detailed info during fitting. Optional.
Default = \code{FALSE}}
}
\value{
Results list object
}
\description{
Perform main fit
}
\examples{
\dontrun{
PerformMainFit(
  context, data, param, info, maxNoFit = 2, ctol = 1e-6, ftol = 1e-5, verbose = TRUE
)
}

}
