% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerformMainFit.R
\name{PerformMainFit}
\alias{PerformMainFit}
\title{PerformMainFit}
\usage{
PerformMainFit(
  context,
  data,
  param = NULL,
  info = NULL,
  maxNoFit = 30L,
  maxRunTime = as.difftime(Inf, units = "secs"),
  ctol = 1e-06,
  ftol = 1e-05,
  algorithm = "NLOPT_LN_BOBYQA",
  attemptSimplify = TRUE,
  verbose
)
}
\arguments{
\item{context}{List of parameters. Required.}

\item{data}{Input data as data.table. Required.}

\item{param}{List of parameters. Optional. Default = NULL.}

\item{info}{List of parameters. Optional. Default = NULL.}

\item{maxNoFit}{Maximum number of amoeba iterations. Optional. Default = 30.}

\item{maxRunTime}{Maximum allowed run time as \code{difftime} object. Fit will be assumed to not
converging when exceeding this time. If NULL then no time out. Optional.
Default = as.difftime(Inf, units = 'secs').}

\item{ctol}{Minium required deviance in consecutive lambda estimations.
Optional. Default = 1e-6.}

\item{ftol}{Minium required deviance in amoeba calculations. Optional.
Default = 1e-5.}

\item{algorithm}{Name of optimization algorithm from package \code{nloptr} to use for bootstrap
iterations. Default = 'NLOPT_LN_BOBYQA'}

\item{attemptSimplify}{Logical indicating to attempting simplifying the model to fit it a more
complex model did not converge. The simplification is achieved by reducing the number of knots
in the spline. Optional. Default = TRUE.}

\item{verbose}{Logical indicating to print detailed info during fitting. Optional. If missing
then the value of \code{context$Settings$Verbose} is used.}
}
\value{
Results list object
}
\description{
Perform main fit
}
\examples{
\dontrun{
PerformMainFit(
  context, data, param, info, maxNoFit = 2, ctol = 1e-6, ftol = 1e-5, verbose = TRUE
)
}

}
