% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerformMainFit.R
\name{PerformMainFit}
\alias{PerformMainFit}
\title{PerformMainFit}
\usage{
PerformMainFit(context)
}
\arguments{
\item{context}{List of parameters. Required.}
}
\value{
context list object
}
\description{
Description
}
\examples{
\dontrun{
PerformMainFit(context)
}

}
