% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreprocessInputData.R
\name{PreprocessInputData}
\alias{PreprocessInputData}
\title{PreprocessInputData}
\usage{
PreprocessInputData(
  inputData,
  minYear = 1980L,
  maxYear = year(Sys.time()),
  verbose = verbose
)
}
\arguments{
\item{inputData}{List of input data. Required.}

\item{minYear}{Model start year.}

\item{maxYear}{Model end year.}

\item{verbose}{Logical indicating to print out messages (TRUE) or not (FALSE)}
}
\value{
Input data as data.table
}
\description{
Description
}
\examples{
\dontrun{
PreprocessInputData(inputData)
}

}
