% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreprocessInputData.R
\name{PreprocessInputData}
\alias{PreprocessInputData}
\title{PreprocessInputData}
\usage{
PreprocessInputData(inputData)
}
\arguments{
\item{inputData}{List of input data. Required.}
}
\value{
Input data as data.table
}
\description{
Description
}
\examples{
\dontrun{
PreprocessInputData(inputData)
}

}
