% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Messages.R
\name{PrintAlert}
\alias{PrintAlert}
\title{PrintAlert}
\usage{
PrintAlert(..., collapse = " ", type = "info", .envir = parent.frame())
}
\arguments{
\item{...}{Text to be printed}

\item{collapse}{String to be used for concatenating texts}

\item{type}{Type of alert}

\item{.envir}{Environment for lookup of variables referenced in the text}
}
\description{
PrintAlert
}
\examples{
PrintAlert('Test')
PrintAlert('Test', type = 'success')
PrintAlert('Test', type = 'danger')
PrintAlert('Test', type = 'warning')
PrintAlert('Test', type = 'info')

}
