% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Messages.R
\name{PrintBullets}
\alias{PrintBullets}
\title{PrintBullets}
\usage{
PrintBullets(items = c(), .envir = parent.frame(), verbose = TRUE)
}
\arguments{
\item{items}{Vector of text items}

\item{.envir}{Environment for lookup of variables referenced in the text}

\item{verbose}{Logical indicating to print out messages (TRUE) or not (FALSE)}
}
\description{
PrintBullets
}
\examples{
\dontrun{
PrintBullets(c('Item 1', 'Item 2'))
}
}
