% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Messages.R
\name{PrintBullets}
\alias{PrintBullets}
\title{PrintBullets}
\usage{
PrintBullets(items = c(), .envir = parent.frame())
}
\arguments{
\item{items}{Vector of text items}

\item{.envir}{Environment for lookup of variables referenced in the text}
}
\description{
PrintBullets
}
\examples{
PrintBullets(c('Item 1', 'Item 2'))

}
