% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Messages.R
\name{PrintH1}
\alias{PrintH1}
\title{PrintH1}
\usage{
PrintH1(..., collapse = " ", verbose = TRUE, .envir = parent.frame())
}
\arguments{
\item{...}{Text to be printed}

\item{collapse}{String to be used for concatenating texts. Default = ' '}

\item{verbose}{Logical indicating to print the message. Default = TRUE.}

\item{.envir}{Environment for lookup of variables referenced in the text}
}
\description{
PrintH1
}
\examples{
PrintH1('Test')
}
