% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Messages.R
\name{PrintH1}
\alias{PrintH1}
\title{PrintH1}
\usage{
PrintH1(..., collapse = " ", .envir = parent.frame())
}
\arguments{
\item{...}{Text to be printed}

\item{collapse}{String to be used for concatenating texts}

\item{.envir}{Environment for lookup of variables referenced in the text}
}
\description{
PrintH1
}
\examples{
PrintH1('Test')

}
