% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadInputData.R
\name{ReadInputData}
\alias{ReadInputData}
\title{ReadInputData}
\usage{
ReadInputData(context)
}
\arguments{
\item{context}{List of parameters. Required.}
}
\value{
NULL (invisibly)
}
\description{
Description
}
\examples{
\dontrun{
ReadInputData(context)
}

}
