% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadInputData.R
\name{ReadInputData}
\alias{ReadInputData}
\title{ReadInputData}
\usage{
ReadInputData(inputDataPath, verbose = TRUE)
}
\arguments{
\item{inputDataPath}{Input data path}

\item{verbose}{Logical indicating to print out messages (TRUE) or not (FALSE)}
}
\value{
Input data as list
}
\description{
Description
}
\examples{
\dontrun{
ReadInputData(inputDataPath)
}

}
