% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadModelFile.R
\name{ReadModelFile}
\alias{ReadModelFile}
\title{ReadModelFile}
\usage{
ReadModelFile(context)
}
\arguments{
\item{context}{List of parameters. Required.}
}
\value{
NULL (invisibly)
}
\description{
Reads XML model file
}
\examples{
\dontrun{
ReadModelFile(context)
}

}
