% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadModelFile.R
\name{ReadModelFile}
\alias{ReadModelFile}
\title{ReadModelFile}
\usage{
ReadModelFile(modelFilePath = NULL, inputDataPath = NULL, verbose = TRUE)
}
\arguments{
\item{modelFilePath}{Model file path. Optional. Default = NULL.}

\item{inputDataPath}{Input data path. Optional. Default = NULL}

\item{verbose}{Logical indicating to print out messages (TRUE) or not (FALSE)}
}
\value{
NULL (invisibly)
}
\description{
Reads XML model file
}
\examples{
\dontrun{
ReadModelFile(context)
}

}
