% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunIncidenceModel.R
\name{RunIncidenceModel}
\alias{RunIncidenceModel}
\title{Run incidence model}
\usage{
RunIncidenceModel(...)
}
\arguments{
\item{...}{Arguments to be passed to \code{\link{GetRunContext}}.}
}
\value{
list object containing context, data and model objects
}
\description{
Executes the incidence model
}
\examples{
RunIncidenceModel()

}
